// src/app/models/fodder.ts
export interface Fodder {
  main?: number;
  main2?: number;
  main3?: number;
  main4?: number;
  main5?: number;
  main6?: number;
  main7?: number;
  main8?: number;
  main9?: number;
  main10?: number;
  main11?: number;
  main12?: number;
  main13?: number;
  main14?: number;
  main15?: number;
  main16?: number;
  main17?: number;
  main18?: number;
  main19?: number;
  main20?: number;
  secondary?: number;
  secondary2?: number;
  secondary3?: number;
  secondary4?: number;
  secondary5?: number;
  secondary6?: number;
  secondary7?: number;
  secondary8?: number;
  secondary9?: number;
  secondary10?: number;
  secondary11?: number;
  secondary12?: number;
  secondary13?: number;
  secondary14?: number;
  secondary15?: number;
  secondary16?: number;
  secondary17?: number;
  secondary18?: number;
  secondary19?: number;
  secondary20?: number;
  fodder?: number;
  sixStar?: number;
  sevenStar?: number;
  eightStar?: number;
  nineStar?: number;
  tenStar?: number;
  e1?: number;
  e2?: number;
}